+++ packages/amplify-sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * feat: Added current `team` to authenticated platform account object.
  * fix: Fixed retrieving list of roles implicitly using the org from the account object.
+ * fix: Updated Axway ID preprod endpoint.
  * chore: Updated dependencies.
 
 # v3.0.2 (Nov 5, 2021)