+++ packages/axway-cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v3.0.3
 
+ * fix: Don't show available updates when `--json` flag is set.
  * chore: Updated dependencies.
 
 # v3.0.2 (Nov 5, 2021)